Revert "Removed commit a61d4143c7f1a8607fec62bb07c3fc1c264e6e4c"

+++ lib/dailyploy/schema/task_list_tasks.ex
@@ -20,7 +20,7 @@ defmodule Dailyploy.Schema.TaskListTasks do
   schema "task_list_tasks" do
     field :name, :string
     field :description, :string
-    field :estimation, :float
+    field :estimation, :integer
     field :status, :string, default: "not_started"
     field :priority, :string, default: "no_priority"
     field :identifier, :string
